// backend/src/app/modules/shared/enums/element-target.enum.ts
export enum ElementTarget {
    PAGES_COUNT = 'PAGES_COUNT',
    PAGE_URL_BY_NUMBER = 'PAGE_URL_BY_NUMBER',
    PAGES_DETAIL_URLS = 'PAGES_DETAIL_URLS',
    ARTICLES_DATA_DELIMITER = 'ARTICLES_DATA_DELIMITER',
    ARTICLE_BRAND = 'ARTICLE_BRAND',
    ARTICLE_DESCRIPTION = 'ARTICLE_DESCRIPTION',
    ARTICLE_PRICE = 'ARTICLE_PRICE',
    ARTICLE_COLOR = 'ARTICLE_COLOR',
    ARTICLE_SIZES = 'ARTICLE_SIZES',
    ARTICLE_IMAGES = 'ARTICLE_IMAGES'
}